backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
firebase-admin>=5.0.0
python-dotenv>=0.19.0
